+++ test/home.js
@@ -5,11 +5,5 @@ import {local_storage_get, local_storage_remove} from '../Storage.js';
 on_page_load(() =>
 {
 	let body = get_body();
-	body.append_child(
-		new Element().text(local_storage_get('username')),
-		new Element().text('Welcome')
-
-	);
-	local_storage_remove('username');
 
 });